codespaces-jupyter-main/BusinessThis/requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
plotly>=5.17.0

# Backend dependencies
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
stripe>=7.0.0
pyjwt>=2.8.0
werkzeug>=2.3.0
gunicorn>=21.2.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Reports and visualization
reportlab>=4.0.0
openpyxl>=3.1.0

# AI and ML
openai>=1.0.0
scikit-learn>=1.3.0

# Email services
sendgrid>=6.10.0

# Bank integration
plaid-python>=9.0.0

# Caching and performance
redis>=4.6.0